requirements.txt
# Core
flask==3.1.*
gunicorn==23.*
pdfplumber==0.11.*
python-docx==1.1.*
requests==2.32.*
python-dotenv==1.1.*
apscheduler==3.10.*
flask-login==0.6.*
flask-wtf==1.2.*
flask-limiter==3.8.*
tenacity==9.*
pytest==8.*

# Intelligence
anthropic==0.49.*
beautifulsoup4==4.12.*
geopy==2.4.*
